Fixed customer in main.c
Customer no longer waits as it should. Added a training_floor array to keep track of the training floor space availability. Trainer needs to be fixed. Comments above functions explaining what they should do. Not sure how to go about trainers being with a customer and checking waiting room upon releasing a customer. Also pthread cust[] should be the customers that are coming in, not the total amount of customers the gym can hold.